Remove exit pkg from async gateway
